Settings reworked
* Builds are now the central paradigm, ie live, empty, test. Database names
are only used when using mysql. Internally speaking, this is a minor change
but practical use has shown it's much more natural and portable to think
in terms of different build processes rather than coupling to database
names
* 'database', 'password' removed. They should now be specified in either
settings or settings.local. This is a simpler system, especially since
settings.local solves the "I can't commit my password" issue
* db-load-settings heavily refactored. It should now be more tolerant in
the face of passed arguments, and should prevent any reversed-alias issues
* db-init now streamlined, so it doesn't test for success on as many
operations. This means a failed init is not as pretty to look at, but I
think it's more indicative of the underlying problem. It's also much easier
to maintain